+++ src/components/charts/KeyFiguresChart.tsx
